+++ tests/test_resolver.py
@@ -10,7 +10,7 @@ import pytest
 
         (['Flask'],
          ['flask==0.10.1', 'itsdangerous==0.24', 'markupsafe==0.23',
-         'jinja2==2.7.3', 'werkzeug==0.10.4']),
+          'jinja2==2.7.3', 'werkzeug==0.10.4']),
 
         (['Jinja2', 'markupsafe'],
          ['jinja2==2.7.3', 'markupsafe==0.23']),
@@ -79,10 +79,18 @@ import pytest
              'librabbitmq==1.6.1',
              'pytz==2016.4']
          ),
+
+        # We shouldn't include irrelevant pip constraints
+        # See: GH-417
+        (['Flask', ('click', True), ('itsdangerous', True)],
+         ['flask==0.10.1', 'itsdangerous==0.24', 'markupsafe==0.23',
+          'jinja2==2.7.3', 'werkzeug==0.10.4']
+         )
     ])
 )
 def test_resolver(resolver, from_line, input, expected, prereleases):
-    input = [from_line(line) for line in input]
+    input = [line if isinstance(line, tuple) else (line, False) for line in input]
+    input = [from_line(req[0], constraint=req[1]) for req in input]
     output = resolver(input, prereleases=prereleases).resolve()
     output = {str(line) for line in output}
     assert output == {str(line) for line in expected}